server/services/recognition_service.py
from fastapi import Depends,UploadFile, File, APIRouter
from PIL import Image
from io import BytesIO



def recognize_image(image: Image.Image) -> int:
    # Burada gerçek bir tanıma işlemi yapılmalıdır.
    # Yapay zeka modelinizden sonucu alın
    # Örnek olarak şu an sadece bir sabit integer dönüyorum.
    return 42

async def process_and_store_image(file: UploadFile = File(...)):
    # Upload edilen resmi işleme ve saklama adımlarını gerçekleştirin
    image = Image.open(BytesIO(file.file.read()))
    
    # Burada işlenmiş görüntüyü saklama veya özel işlemleri gerçekleştirme adımlarını ekleyebilirsiniz
    # Şu an sadece orijinal resmi döndürüyorum
    processed_image = image

    return processed_image
